+++ tests/integration/dedicated_alb/provision.py
@@ -1,0 +1,44 @@
+from broker.extensions import db
+from broker.models import (
+    DedicatedALB,
+    DedicatedALBListener,
+)
+
+
+def create_dedicated_alb_listeners(organization_guid, dedicated_alb_id):
+    alb_0 = DedicatedALB(
+        alb_arn="alb-our-arn-0",
+        dedicated_org=organization_guid,
+        id=dedicated_alb_id,
+    )
+    alb_1 = DedicatedALB(
+        alb_arn="alb-our-arn-1",
+        dedicated_org=organization_guid,
+    )
+    db.session.add_all(
+        [
+            alb_0,
+            alb_1,
+        ]
+    )
+    db.session.commit()
+
+    our_listener_0 = DedicatedALBListener(
+        listener_arn="our-arn-0",
+        alb_arn="alb-our-arn-0",
+        dedicated_org=organization_guid,
+    )
+    our_listener_1 = DedicatedALBListener(
+        listener_arn="our-arn-1",
+        alb_arn="alb-our-arn-1",
+        dedicated_org=organization_guid,
+    )
+
+    db.session.add_all(
+        [
+            our_listener_0,
+            our_listener_1,
+        ]
+    )
+    db.session.commit()
+    db.session.expunge_all()
